✨ Create "My submissions" page

Closes #405

See merge request karaokemugen/code/karaokemugen-server!119

+++ kmserver-core/src/services/inbox.ts
@@ -1,30 +1,31 @@
 import { promises as fs } from 'fs';
+import i18n from 'i18next';
 import { resolve } from 'path';
 import { v4 as uuidV4 } from 'uuid';
 
-import {clearInbox, deleteInbox, insertInbox, selectInbox, updateInboxDownloaded, updateInboxUnassign} from '../dao/inbox.js';
+import {deleteInbox, insertInbox, selectInbox, updateInboxDownloaded, updateInboxStatus, updateInboxUnassign} from '../dao/inbox.js';
 import { deleteKara } from '../dao/kara.js';
 import { refreshAllKaraTag } from '../dao/tag.js';
 import { formatKaraV4, getDataFromKaraFile } from '../lib/dao/karafile.js';
 import { getDataFromTagFile } from '../lib/dao/tagfile.js';
-import { readAllKaras } from '../lib/services/generation.js';
 import { refreshKarasAfterDBChange } from '../lib/services/karaManagement.js';
 import {KaraMetaFile, MetaFile, TagMetaFile} from '../lib/types/downloads.js';
-import { DBInbox, Inbox } from '../lib/types/inbox.js';
+import { DBInbox, Inbox, InboxActions } from '../lib/types/inbox.js';
 import { KaraFileV4 } from '../lib/types/kara.js';
 import { TagFile } from '../lib/types/tag.js';
+import { JWTTokenWithRoles } from '../lib/types/user.js';
 import { getConfig, resolvedPathRepos } from '../lib/utils/config.js';
 import { tagTypes } from '../lib/utils/constants.js';
 import { ErrorKM } from '../lib/utils/error.js';
-import { fileExists, listAllFiles } from '../lib/utils/files.js';
-import { assignIssue, closeIssue } from '../lib/utils/gitlab.js';
+import { assignIssue, closeIssue, postNoteToIssue } from '../lib/utils/gitlab.js';
 import logger from '../lib/utils/logger.js';
-import { findFileByUUID } from '../utils/files.js';
+import { adminToken } from '../utils/constants.js';
+import { sendMail } from '../utils/mailer.js';
 import sentry from '../utils/sentry.js';
 import { getKara } from './kara.js';
 import { getRepos } from './repo.js';
 import { getTag } from './tag.js';
-import { findUserByName } from './user.js';
+import { findUserByName, getUserLanguage } from './user.js';
 
 const service = 'Inbox';
 
@@ -78,20 +79,100 @@ export async function getKaraInbox(inid: string): Promise<Inbox> {
 	}
 }
 
-export async function getInbox(isMaintainer: boolean): Promise<DBInbox[]> {
-	const listInbox = await selectInbox();
-	if (!isMaintainer) listInbox.forEach(inbox => {
-		delete inbox.contact;
-		delete inbox.fk_login;
-	});
+export async function getInbox(isMaintainer: boolean, byUser?: string): Promise<DBInbox[]> {
+	const listInbox = await selectInbox(undefined, byUser);
+	if (!isMaintainer) listInbox.forEach(inbox => delete inbox.contact);
 	return listInbox;
+}
+
+export async function setInboxStatus(inid: string, status: InboxActions, reason?: string) {
+	try {
+		const inbox = (await selectInbox(inid))[0];
+		if (!inbox) throw new ErrorKM('INBOX_UNKNOWN_ERROR', 404, false);
+		if (!Array.isArray(inbox.history)) inbox.history = [];
+		inbox.history.push({
+			action: status,
+			details: reason,
+			datetime: new Date()
+		});
+		await updateInboxStatus(inid, status, inbox.history, status === 'rejected' ? reason : null);
+
+		// What to do with the different statuses
+		const user = await findUserByName(inbox.username, { contact: true });
+		const repoName = getConfig().System.Repositories[0].Name;
+
+		if (status === 'changes_requested') {
+			if (user.flag_contributor_emails) await sendMail(
+				i18n.t('MAIL.INBOX.CHANGES_REQUESTED.SUBJECT', { lng: getUserLanguage(user) }),
+				i18n.t('MAIL.INBOX.CHANGES_REQUESTED.BODY', {
+					lng: getUserLanguage(user),
+					reason,
+				}),
+				user.login,
+				user.email
+			);
+			if (inbox.gitlab_issue) {
+				const issueNumber = getGitlabIssueNumber(inbox.gitlab_issue);
+				await postNoteToIssue(issueNumber, repoName, `Waiting for changes by original uploader. Notes from reviewer:
+				${reason}
+				`);
+			}
+		} else if (status === 'in_review') {
+			if (user.flag_contributor_emails) await sendMail(
+				i18n.t('MAIL.INBOX.IN_REVIEW.SUBJECT', { lng: getUserLanguage(user) }),
+				i18n.t('MAIL.INBOX.IN_REVIEW.BODY', {
+					lng: getUserLanguage(user)
+				}),
+				user.login,
+				user.email
+			);
+			if (inbox.gitlab_issue) {
+				const issueNumber = getGitlabIssueNumber(inbox.gitlab_issue);
+				await postNoteToIssue(issueNumber, repoName, `Upload is being reviewed by ${inbox.username_downloaded}`);
+			}
+		} else if (status === 'accepted') {
+			if (user.flag_contributor_emails) await sendMail(
+				i18n.t('MAIL.INBOX.ACCEPTED.SUBJECT', { lng: getUserLanguage(user) }),
+				i18n.t('MAIL.INBOX.ACCEPTED.BODY', {
+					lng: getUserLanguage(user)
+				}),
+				user.login,
+				user.email
+			);
+			if (inbox.gitlab_issue) {
+				const issueNumber = getGitlabIssueNumber(inbox.gitlab_issue);
+				await postNoteToIssue(issueNumber, repoName, `Upload was ACCEPTED by ${inbox.username_downloaded}`);
+				await closeIssue(issueNumber, repoName);
+			}
+		} else if (status === 'rejected') {
+			if (user.flag_contributor_emails) await sendMail(
+				i18n.t('MAIL.INBOX.REJECTED.SUBJECT', { lng: getUserLanguage(user) }),
+				i18n.t('MAIL.INBOX.REJECTED.BODY', {
+					lng: getUserLanguage(user),
+					reason,
+				}),
+				user.login,
+				user.email
+			);
+			if (inbox.gitlab_issue) {
+				const issueNumber = getGitlabIssueNumber(inbox.gitlab_issue);
+				await postNoteToIssue(issueNumber, repoName, `Upload was REJECTED by ${inbox.username_downloaded}: ${reason}`);
+				await closeIssue(issueNumber, repoName);
+			}
+		}
+	} catch (err) {
+		logger.error(`Failed to set inbox item ${inid} status to ${status}`, {service, obj: err});
+		sentry.error(err);
+		throw err instanceof ErrorKM ? err : new ErrorKM('SET_INBOX_STATUS_ERROR');
+	}
 }
 
 export async function markKaraInboxAsDownloaded(inid: string, username: string) {
 	try {
 		const inbox = (await selectInbox(inid))[0];
 		if (!inbox) throw new ErrorKM('INBOX_UNKNOWN_ERROR', 404, false);
-		return await updateInboxDownloaded(username, inid);
+		await updateInboxDownloaded(username, inid);
+		await setInboxStatus(inid, 'in_review');
 	} catch (err) {
 		logger.error(`Failed to mark inbox item ${inid} as downloaded by ${username}`, {service, obj: err});
 		sentry.error(err);
@@ -99,7 +180,7 @@ export async function markKaraInboxAsDownloaded(inid: string, username: string) 
 	}
 }
 
-function getGitlabIssueNumber(url: string): number {
+export function getGitlabIssueNumber(url: string): number {
 	return +url.split('/')[url.split('/').length - 1];
 }
 
@@ -121,79 +202,60 @@ export async function markKaraInboxAsUnassigned(inid: string, username: string) 
 }
 
 
-export async function addKaraInInbox(kara: KaraFileV4, contact: {name: string, login?: string}, issue?: string, edited_kid?: string) {
+export async function addKaraInInbox(kara: KaraFileV4, contact: {name: string, login?: string}, issue?: string, edited_kid?: string, inid?: string) {
 	try {
+		inid = inid || uuidV4();
 		await insertInbox({
-			inid: uuidV4(),
+			inid,
 			name: kara.data.songname,
 			created_at: new Date(),
 			gitlab_issue: issue,
 			contact: contact.name,
 			kid: kara.data.kid,
 			edited_kid,
-			fk_login: contact.login ?? null
+			username: contact.login ?? null,
+			mediafile: kara.medias[0].filename
 		});
+		return inid;
 	} catch (err) {
 		logger.error('Unable to create kara in inbox', {service, obj: err});
 		sentry.error(err);
 	}
 }
 
-export async function removeProcessedInboxes() {
-	logger.info('Removing possible processed inbox items if they are present in main repository', { service });
-	const inbox = await getInbox(false);
-	// Get a list of KIDs from the main repository (not including Staging then)
-	const karaFiles = await listAllFiles('Karaokes', getConfig().System.Repositories[0].Name);
-	const karas = await readAllKaras(karaFiles, false);
-	const kids = new Set(karas.map(k => k.data.kid));
-	for (const inboxItem of inbox) {
-		if (kids.has(inboxItem.kid)) {
-			await removeKaraFromInbox(inboxItem.inid);
-		}
-	}
-}
-
-export async function removeKaraFromInbox(inid: string) {
+export async function removeKaraFromInbox(inid: string, authToken: JWTTokenWithRoles) {
 	try {
 		const inbox = (await selectInbox(inid))[0];
 		if (!inbox) throw new ErrorKM('INBOX_UNKNOWN_ERROR', 404, false);
+		if (!authToken.roles.admin && !authToken.roles.maintainer) {
+			if (inbox.username !== authToken.username) throw new ErrorKM('INBOX_DELETE_FORBIDDEN', 403, false);
+		}
 		// Kara might not exist anymore if something went wrong.
 		try {
 			const kara = await getKara({
 				q: `k:${inbox.edited_kid || inbox.kid}!r:Staging`
 			});
-			// If kara is not found, it means the song has already been added to the database
+			// If kara is not found, it means the song has already been added to the database or it's not there yet because the user decides to remove their inbox shortly after submitting it
 			if (kara) {
 				const karaData = formatKaraV4(kara);
 				await deleteKara([kara.kid]);
-				await refreshKarasAfterDBChange('DELETE', [karaData.data]);
-				await refreshAllKaraTag();
+				refreshKarasAfterDBChange('DELETE', [karaData.data]);
+				refreshAllKaraTag();
 			}
 		} catch (err) {
 			// Non-fatal.
 			logger.info(`Kara ${inbox.name} is not found, it means the song has already been added to the database`, { service });
 		}
-		if (inbox.karafile) {
-			const karaPath = resolve(resolvedPathRepos('Karaokes', 'Staging')[0], inbox.karafile);
-			let subPath: string;
-			if (inbox.lyrics_infos[0]) subPath = resolve(resolvedPathRepos('Lyrics', 'Staging')[0], inbox.lyrics_infos[0].filename);
-			const mediaPath = resolve(resolvedPathRepos('Medias', 'Staging')[0], inbox.mediafile);
-			try {
-				await Promise.all([
-					fs.unlink(karaPath),
-					fs.unlink(mediaPath)
-				]);
-				if (subPath) await fs.unlink(subPath);
-			} catch (err) {
-				// Non-fatal, files might not exist anymore.
-				logger.warn(`Unable to remove some files after inbox deletion for song ${inbox.name}`, { service });
-			}
-		}
-		await deleteInbox(inid);
+		const promises = [];
+		promises.push(removeInboxFiles(inbox.edited_kid || inbox.kid, inbox.karafile, inbox.mediafile, inbox.lyrics_infos ? inbox.lyrics_infos[0]?.filename : null));
+		promises.push(deleteInbox(inid));
 		if (inbox.gitlab_issue) {
 			const numberIssue = getGitlabIssueNumber(inbox.gitlab_issue);
-			await closeIssue(numberIssue, getConfig().System.Repositories[0].Name);
+			const repoName = getConfig().System.Repositories[0].Name;
+			if (inbox.username === authToken.username) promises.push(postNoteToIssue(numberIssue, repoName, 'Closing by request of original uploader'));
+			promises.push(closeIssue(numberIssue, repoName));
 		}
+		await Promise.all(promises);
 	} catch (err) {
 		logger.error(`Failed to delete inbox item ${inid}`, {service, obj: err});
 		sentry.error(err);
@@ -233,40 +295,64 @@ export async function clearUnusedStagingTags() {
 	}
 }
 
-/** Clear inbox of songs that have been validated and added to the main database
- * These songs should have the same KID but not the Staging repository in database.
- * :uuid
- */
-export async function clearOldInboxEntries() {
-	logger.debug('Clearing old inbox entries', {service});
-	const deletedKaras = await clearInbox();
-	for (const kara of deletedKaras) {
-		logger.debug(`${kara.kid} was cleared`, {service, obj: kara});
-		try {
-			// Find and delete files
-			const karaPath = resolve(resolvedPathRepos('Karaokes', 'Staging')[0], kara.karafile);
-			if (await fileExists(karaPath)) {
-				const subPath = resolve(resolvedPathRepos('Lyrics', 'Staging')[0], kara.lyrics_infos[0].filename);
-				const mediaPath = resolve(resolvedPathRepos('Medias', 'Staging')[0], kara.mediafile);
-				await Promise.all([
-					fs.unlink(karaPath),
-					fs.unlink(subPath),
-					fs.unlink(mediaPath)
-				]);
-			} else {
-				const [name, content] = await findFileByUUID('kid', kara.kid, 'Staging');
-				const newKaraPath = resolve(resolvedPathRepos('Karaokes', 'Staging')[0], name);
-				const subPath = resolve(resolvedPathRepos('Lyrics', 'Staging')[0], content.medias[0].lyrics[0].filename);
-				const mediaPath = resolve(resolvedPathRepos('Medias', 'Staging')[0], content.medias[0].filename);
-				await Promise.all([
-					fs.unlink(newKaraPath),
-					fs.unlink(subPath),
-					fs.unlink(mediaPath)
-				]);
-			}
-		} catch (err) {
-			logger.error(`Error when cleaning kara (${kara.kid})`, {service, obj: err});
-			throw err;
+/** Clear inbox entries that have been refused, accepted, or in change_requested status and older than XX days */
+export async function clearInactiveInboxEntries() {
+	const conf = getConfig();
+	const inbox = await getInbox(true);
+	const deletedDate = new Date();
+	deletedDate.setDate(deletedDate.getDate() - conf.Frontend.Import.CleanupDays);
+	const inboxesToClear = inbox
+		.filter(i => i.status === 'accepted' || i.status === 'rejected' || i.status === 'changes_requested')
+		.filter(i => new Date(i.modified_at) < deletedDate);
+	for (const inbox of inboxesToClear) {
+		removeKaraFromInbox(inbox.inid, adminToken);
+	}
+}
+
+/** clear inboxes that have been processed (added to the main repo's database) but not removed
+ * This one is run after generation so it compares listed files used by generation.
+*/
+export async function clearProcessedInboxes(karas: KaraFileV4[]) {
+	logger.info('Removing possible processed inbox items if they are present in main repository', { service });
+	const inbox = await getInbox(false);
+	// Get a list of KIDs from the main repository (not including Staging then)
+	const kids = new Set(karas
+		.filter(k => k.data.repository !== 'Staging')
+		.map(k => k.data.kid));
+	for (const inboxItem of inbox) {
+		if (kids.has(inboxItem.kid)) {
+			await removeKaraFromInbox(inboxItem.inid, adminToken);
 		}
 	}
 }
+
+/** Physically remove inbox files */
+async function removeInboxFiles(kid: string, karaFile?: string, mediaFile?: string, lyricsFile?: string) {
+	try {
+		// Determine if we need to guess which files to remove
+		const mediaDir = resolvedPathRepos('Medias', 'Staging')[0];
+		const lyricsDir = resolvedPathRepos('Lyrics', 'Staging')[0]
+		if (!karaFile || !mediaFile || !lyricsFile) {
+			karaFile = `${kid}.kara.json`;
+			// This is not multi-track drifting-proof
+			const medias = await fs.readdir(mediaDir);
+			mediaFile = medias.filter(f => f.startsWith(kid))[0];
+			const lyrics = await fs.readdir(lyricsDir);
+			lyricsFile = lyrics.filter(f => f.startsWith(kid))[0];
+		}
+		// Find and delete files
+		const karaPath = resolve(resolvedPathRepos('Karaokes', 'Staging')[0], karaFile);
+		const subPath = resolve(lyricsDir, lyricsFile || '');
+		const mediaPath = resolve(mediaDir, mediaFile);
+		const promises = [
+			fs.unlink(karaPath),
+			fs.unlink(mediaPath)
+		]
+		if (lyricsFile) promises.push(fs.unlink(subPath));
+		await Promise.all(promises);
+	} catch (err) {
+		logger.error(`Error when cleaning kara (${karaFile})`, {service, obj: err});
+		// Non-fatal
+	}
+}
+
